+++ model/pack_logo.go
@@ -16,8 +16,8 @@ type PackLogos []*PackLogo
 
 // PackLogo represents a pack logo model definition.
 type PackLogo struct {
-	ID          int              `json:"-" gorm:"primary_key"`
-	PackID      int              `json:"-" sql:"index"`
+	ID          int64            `json:"-" gorm:"primary_key"`
+	PackID      int64            `json:"-" sql:"index"`
 	Pack        *Pack            `json:"-"`
 	Slug        string           `json:"slug" sql:"unique_index"`
 	ContentType string           `json:"content_type"`
